Creacion de modales para funcionar en ambiente y facturaciones y devoluciones implementado
